task 0 fix

+++ python-server_side_rendering/task_00_intro.py
@@ -5,8 +5,8 @@ import os
 def generate_invitations(template_content, attendees):
     x = 1
     for attendee in attendees:
-        #if not (os.path.exists(f"output_{x}.txt")):
-        #    return
+        if not (os.path.exists(f"output_{x}.txt")):
+            return
         with open(f"output_{x}.txt", 'w') as f:
             try:
                 invitation = template_content.format(**attendee)
